moved DirectusProvider into base folder
