+++ jest.config.json
@@ -5,6 +5,7 @@
   "testMatch": ["<rootDir>/src/**/*(*.)@(spec|test).[tj]s?(x)"],
   "moduleNameMapper": {
     "^@/(.*)$": "<rootDir>/src/$1",
-    "^@components/(.*)$": "<rootDir>/src/components/$1"
+    "^@components/(.*)$": "<rootDir>/src/components/$1",
+    "^@base/(.*)$": "<rootDir>/src/base/$1"
   }
 }
