+++ tsconfig.json
@@ -15,6 +15,7 @@
     "paths": {
       "@/*": ["*"],
       "@components/*": ["components/*"],
+      "@base/*": ["base/*"],
     }
   }
 }
